+++ docs/source/whatsnew.rst
@@ -5,6 +5,7 @@ What's New
 
 v4.3.0
 ------
+- DEPRECATION: :meth:`Person.add_device()` has been deprecated. Use :meth:`Person.devices.create()` instead
 - BREAKING CHANGE: :attr:`Org.numbers` and :attr:`Location.numbers` now uses :class:`NumberList`
 - :meth:`AutoAttendantList.create()` now supports ``business_hours_menu`` and ``after_hours_menu`` as optional arguments, building an empty menu when omitted
 - The :class:`Number` and :class:`NumberList` classes have been added to handle all number-management and search functions